3-code/kanban-sync/pyproject.toml
[project]
name = "kanban-sync"
version = "0.0.1"
description = "Obsidian Kanban file I/O + sync-vs-edit boundary detection for project-agent-system."
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "pydantic>=2.9,<3",
    "canonical-json",
    "bearer-auth",
    "pyyaml>=6.0.3",
    "pydantic-settings>=2.14.0",
    "httpx>=0.27.2",
    "filelock>=3.29.0",
]

[tool.uv.sources]
canonical-json = { path = "../_common/canonical_json", editable = true }
bearer-auth = { path = "../_common/bearer_auth", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<0.25",
    "httpx>=0.27,<0.28",
    "ruff>=0.7,<0.8",
    "mypy>=1.13,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
    "SIM",
    "TCH",
    "RUF",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
